amélioration pour rendre le jeu logique

+++ src/main/java/model/Doctor.java
@@ -4,10 +4,12 @@ import util.Position;
 import util.TargetStrategy;
 
 import java.util.List;
+import java.util.Random;
 import java.util.Set;
 
 public class Doctor extends AbstractAgent {
     private final TargetStrategy strategy = new TargetStrategy();
+    Random randomNumbers=new Random();
 
     public Doctor(Position position) {
         super(position);
@@ -16,6 +18,7 @@ public class Doctor extends AbstractAgent {
     @Override
     public void update(BoardContext context) {
         Set<Position> targets = context.getPositions(ModelElement.VIRUS);
+        Position lastPos=this.position;
         Position targetPosition = strategy.neighborClosestToFire(
                 this.position,
                 targets,
@@ -27,8 +30,9 @@ public class Doctor extends AbstractAgent {
 
         List<ModelElement> contentOnCase = context.getState(this.position);
 
-        if (contentOnCase.contains(ModelElement.VIRUS)) {
+        if (contentOnCase.contains(ModelElement.VIRUS) && randomNumbers.nextDouble()<0.5 ) {
             context.kill(ModelElement.VIRUS, this.position);
+            context.spawn(ModelElement.DOCTOR,lastPos);
 
         }
     }
